Basic turtle
This commit adds a little to the turtle functionality, allowing users to
create/destroy a turtle and tell it to move forward (which may not
actually be forward from the player's perspective-- I'm still working on
this).  I've also commented out some of the duplicate code in
HelloListener-- I think the entire file can be deleted, but I haven't
tested that yet.  I'll remove it when I'm sure it won't break anything.