Yes, that fixed it.
I think the issue was that setBlock already existed in CanaryMod, and so
it didn't like ours.  The exception it threw made the rest of the
commands not register, I think.  I changed our sb/setBlock to
sbt/setBlockType, and it seems to work now.  I also updated the command
descriptions to be a little clearer.