Messing with build file
Well, I figured out that some change to the build file was what was my
problem before.  I haven't figured out exactly what yet, but here is a
working version.  Going forward, going up, and turning are working, and
I'm not getting exceptions anymore.  Going backward and down are not
working, so that still needs work.  Also, we may want to consider
switching the order that moving/block placing occur in, like we
discussed, but that's not really a bug.

I'm going to keep messing with this to see what the problem was.