Finished undo, invoke (formerly execute)
It turned out there was already a Canary command called execute, so ours
is now invoke.  So there's now a data structure to store scripts and we
can invoke and undo those stored in it.  We still need to update the
hook handler to put things in the ScriptManager, though.