Added support for Turtle3D as well as Turtle3DBase
Turtle3D uses a main() method style rather than inheritance like
Turtle3DBase. Had to use the currently running thread as the key to
prevent race conditions (i.e. if two people on the server at the same
time create turtles, how do we know which one is which when we put them
into the map? answer is to check which thread created each Turtle). Sort
of a hack: We decide whether code is Turtle3D or Turtle3DBase by looking
for the full path to Turtle3DBase in the text of the code someplace
(probably the import statement).

JSONUtil now handles Boolean values.