Added Canarymod boilerplate license to derived DB code
Need to reprint the license boilerplate from Canarymod

Copyright (c) 2012 - 2015, CanaryMod Team
Under the management of PlayBlack and Visual Illusions Entertainment
All rights reserved.

Redistribution and use in source and binary forms, with or without
modification, are permitted provided that the following conditions are
met:
    * Redistributions of source code must retain the above copyright
      notice, this list of conditions and the following disclaimer.
    * Redistributions in binary form must reproduce the above copyright
      notice, this list of conditions and the following disclaimer in
the
      documentation and/or other materials provided with the
distribution.
    * Neither the name of the CanaryMod Team nor the
      names of its contributors may be used to endorse or promote
products
      derived from this software without specific prior written
permission.

THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
IS" AND
ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED
WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
DISCLAIMED. IN NO EVENT SHALL CANARYMOD TEAM OR ITS CONTRIBUTORS BE
LIABLE FOR ANY
DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
DAMAGES
(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
SERVICES;
LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
AND
ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR
TORT
(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
THIS
SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Any source code from the Minecraft Server is not owned by CanaryMod
Team, PlayBlack,
Visual Illusions Entertainment, or its contributors and is not covered
by above license.
Usage of source code from the Minecraft Server is subject to the
Minecraft End User License Agreement as set forth by Mojang AB.
The Minecraft EULA can be viewed at
https://account.mojang.com/documents/minecraft_eula
CanaryMod Team, PlayBlack, Visual Illusions Entertainment, CanaryLib,
CanaryMod, and its contributors
are NOT affiliated with, endorsed, or sponsored by Mojang AB, makers of
Minecraft.
"Minecraft" is a trademark of Notch Development AB
"CanaryMod" name is used with permission from FallenMoonNetwork.